angrylion: do the matrix transform in RSP fixed point, not float
The HLE frontend stored and multiplied its matrices in float and transformed
vertices with float arithmetic. The RSP does all of this in its fixed-point
vector unit, so the float path could never match the LLE pipeline bit-for-bit:
it produced "close but slightly off" coordinates and depended on float rounding
that diverges from the hardware. This is the first step of moving the transform
onto the RSP's actual fixed-point representation.

Matrices are now stored as s15.16 (the integer s16 half in the high 16 bits,
the 1/65536 fractional u16 half in the low 16 -- exactly the two halves the RSP
keeps, concatenated). load_n64_matrix reads the two halves and concatenates
them with no float conversion. mtx_mul multiplies two s15.16 matrices by
accumulating each column's products at 64-bit width and truncating back to
s15.16 (>> 16), which reproduces the "middle 32 bits retained" behaviour of the
RSP's 48-bit vector accumulator for the matrix concatenation. The per-vertex
transform multiplies the integer s16 model position by the s15.16 combined
matrix, accumulates at 64-bit width, and reads back the s15.16 clip
coordinates.

The perspective divide is still float for now (the fixed-point reciprocal,
which uses the RSP's div_ROM table, is a separate follow-up); only the matrix
and transform arithmetic changes here. Verified against the cxd4 LLE reference:
the emitted triangle screen extents and z (z >> 16 ~ 31790 vs the reference's
~31750) track the reference, and the 2D/HUD frames are unchanged. Round-trip
checksum 42896360 preserved.

ISO C89 / MSVC-compatible.

+++ mupen64plus-video-angrylion/rdp_emit_frontend.h
@@ -31,10 +31,15 @@ typedef struct GSPVertex
 
 typedef struct GSPState
 {
-    float projection[4][4];
-    float modelview[GSP_MTX_STACK][4][4];
+    /* Matrices are stored in the RSP's fixed-point form: each element is an
+     * s15.16 value held in an int32 (integer part in the high 16 bits, the
+     * 1/65536 fraction in the low 16). The transform and multiply accumulate
+     * these the way the RSP vector unit does, rather than in float, so the
+     * emitted coordinates match the LLE path bit-for-bit. */
+    int32_t projection[4][4];
+    int32_t modelview[GSP_MTX_STACK][4][4];
     int   modelview_top;
-    float combined[4][4];
+    int32_t combined[4][4];
     int   combined_valid;
 
     BridgeViewport viewport;
